requirements.txt
amqp==5.1.1
APScheduler==3.6.3
arrow==1.2.3
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.5
cycler==0.11.0
Deprecated==1.2.13
fonttools==4.37.3
idna==3.4
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.5.3
numpy==1.23.3
packaging==21.3
paho-mqtt==1.6.1
peewee==3.15.3
Pillow==9.2.0
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
six==1.16.0
tornado==6.1
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
